Refactor Featured Game Component
